Changed/Added some methods
Added a return to the addItem method that returns the itemID if the inventory isn't full so that the player will know what kind of item is in the inventory and what needs to happen afterwards. Also added a new testing application method so the inventory functions can be tested stand-alone with the main file through the command-line.