Game compiles but tile rendering is inaccurate
Added conditionals to render tiles only if they have already been visited or are within 1 tile in any direction of player (playerPosX, playerPosY) but the Map is not rendering as expected. Just committing for now, will check in with the rendering methods author to figure out how to patch soon. Code is ready to be implemented that allows the shader to render 2 tiles in any direction if binoculars are owned, however the Items class must be made abstract first,